dapgd/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0

# EDM dependencies (check EDM's requirements.txt)
ninja>=1.11.0  # For CUDA compilation
psutil>=5.9.0

# Data and I/O
h5py>=3.8.0  # For HDF5 files
imageio>=2.28.0  # Image I/O
tifffile>=2023.4.12  # For TIFF (scientific imaging)
rawpy>=0.18.0  # For RAW camera files
astropy>=5.2.0  # For FITS (astronomy)

# Evaluation and metrics
scikit-image>=0.20.0  # For SSIM and other metrics
lpips>=0.1.4  # Perceptual similarity

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking
omegaconf>=2.3.0  # Hierarchical configs

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0  # Parallel test execution

# Development
black>=23.3.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.3.0  # Type checking
ipython>=8.14.0  # Interactive shell
jupyter>=1.0.0  # Notebooks

# Utilities
tqdm>=4.65.0  # Progress bars
click>=8.1.0  # CLI interface
matplotlib>=3.7.0  # Plotting
